refactor(agent): extract message-metadata helpers (#3250)
Second incremental, behavior-preserving extraction from agent.store.ts
(#3250). Moves the pure DB-row <-> AgentMessage (de)serialization seam —
parseAgentConversationMetadata, serializeAgentConversationMetadata,
parsePersistedBlockMetadata, serializeAgentMessageMetadata,
reconstructStoredAgentMessage, isPairedHandoffMetadata, and the
AgentConversationMetadata / PersistedBlockMetadata types — into a
co-located module at src/lib/agent/message-metadata.ts.

These functions are pure — they operate only on their arguments, JSON, and
imported types, with no closure over store state. AgentMessage is type-only
back-imported from the store (erased at runtime, so no import cycle). The
store keeps every call site unchanged, drops the now-unused StoredMessage
import, and re-exports serializeAgentMessageMetadata and
reconstructStoredAgentMessage so the public surface of @/stores/agent.store
is identical. agent.store.ts drops from 9474 to 9346 lines.

No behavior, IPC-contract, or persistence-schema change.

Validation: full unit suite green (2714 passed); the #3247 pinned test
agent-history-persistence (which round-trips these helpers and asserts on
loadPersistedAgentHistory source text) unchanged and passing; Biome clean;
both files type-clean under tsc.

Taariq Lewis, SerenAI, Paloma, and Volume at https://serendb.com
Email: hello@serendb.com
